+++ pcdet/models/backbones_3d/spconv_unet.py
@@ -3,7 +3,7 @@ from functools import partial
 import torch
 import torch.nn as nn
 
-from ...spconv_utils import replace_feature, spconv
+from ...utils.spconv_utils import replace_feature, spconv
 from ...utils import common_utils
 from .spconv_backbone import post_act_block
