+++ pcdet/models/roi_heads/partA2_head.py
@@ -3,7 +3,7 @@ import torch
 import torch.nn as nn
 
 from ...ops.roiaware_pool3d import roiaware_pool3d_utils
-from ...spconv_utils import spconv
+from ...utils.spconv_utils import spconv
 from .roi_head_template import RoIHeadTemplate
